Share functionality added

+++ app/src/main/java/com/example/taskgather/models/UserModel.kt
@@ -4,7 +4,7 @@ class UserModel (
     var userId: String,
     var userName: String,
     var userEmail: String,
-    var totalNotes:Int,
-    var totalImages:Int ) {
+    var totalNotes:Long,
+    var totalImages:Long ) {
     constructor() : this("","","",0,0)
 }
